refactor: move molecule_pairs_opt.py to core/data/
