+++ simba/core/chemistry/mces/mces_computation.py
@@ -13,9 +13,9 @@ from tqdm import tqdm
 from simba.core.chemistry.edit_distance import edit_distance
 from simba.core.chemistry.mces_loader.load_mces import LoadMCES
 from simba.core.data.molecular_pairs import MolecularPairsSet
+from simba.core.data.molecule_pairs_opt import MoleculePairsOpt
 from simba.core.data.spectrum import SpectrumExt
 from simba.logger_setup import logger
-from simba.molecule_pairs_opt import MoleculePairsOpt
 from simba.train_utils import TrainUtils
